# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/species": {
      "get": {
        "operationId": "SpeciesController_getAllSpecies",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get all species",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeciesListResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Species"]
      },
      "post": {
        "operationId": "SpeciesController_createSpecies",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpeciesRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create species",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SpeciesResponseDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Species"]
      }
    },
    "/api/species/{id}": {
      "get": {
        "operationId": "SpeciesController_getSpeciesById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "Get species by ID",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SpeciesResponseDto" }
              }
            }
          },
          "401": { "description": "Unauthorized - you need to be signed in" },
          "404": { "description": "Species not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Species"]
      },
      "put": {
        "operationId": "SpeciesController_updateSpeciesById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpeciesRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update species",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SpeciesResponseDto" }
              }
            }
          },
          "400": { "description": "Invalid input data" },
          "404": { "description": "Species not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Species"]
      },
      "delete": {
        "operationId": "SpeciesController_deleteSpeciesById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "204": { "description": "Delete species" },
          "404": { "description": "Species not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Species"]
      }
    },
...
    "/api/substrates": {
      "get": {
        "operationId": "SubstrateController_getAllSubstrates",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get all substrates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubstrateListResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Substrates"]
      },
      "post": {
        "operationId": "SubstrateController_createSubstrate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubstrateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create substrates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubstrateDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Substrates"]
      }
    },
    "/api/substrates/{id}": {
      "get": {
        "operationId": "SubstrateController_getSubstrateById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "Get substrates by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubstrateDetailResponseDto"
                }
              }
            }
          },
          "401": { "description": "Unauthorized - you need to be signed in" },
          "404": { "description": "Substrates not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Substrates"]
      },
      "put": {
        "operationId": "SubstrateController_updateSubstrateById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubstrateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update substrates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubstrateDetailResponseDto"
                }
              }
            }
          },
          "400": { "description": "Invalid input data" },
          "404": { "description": "Substrates not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Substrates"]
      },
      "delete": {
        "operationId": "SubstrateController_deleteSubstrateById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "204": { "description": "Delete substrates" },
          "404": { "description": "Substrates not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Substrates"]
      }
    },
    "/api/users": {
      "get": {
        "operationId": "UserController_getAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get all users",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserListResponseDto" }
              }
            }
          },
          "401": { "description": "Unauthorized - you need to be signed in" },
          "403": { "description": "Forbidden" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Users"]
      },
      "post": {
        "operationId": "UserController_registerUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Register",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponseDto" }
              }
            }
          },
          "400": { "description": "Invalid input data" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Users"]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UserController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "example": "me", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Get user by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserResponseDto"
                }
              }
            }
          },
          "401": { "description": "Unauthorized - you need to be signed in" },
          "404": { "description": "User not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Users"]
      },
      "put": {
        "operationId": "UserController_updateUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "example": "me", "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserRequestDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update user by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserResponseDto"
                }
              }
            }
          },
          "400": { "description": "Invalid input data" },
          "401": { "description": "Unauthorized - you need to be signed in" },
          "404": { "description": "User not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Users"]
      },
      "delete": {
        "operationId": "UserController_deleteUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "example": "me", "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Delete user" },
          "401": { "description": "Unauthorized - you need to be signed in" },
          "404": { "description": "User not found" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Users"]
      }
    },
    "/api/users/{id}/substrate-assignments": {
      "get": {
        "operationId": "UserController_getSubstrateAssignments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "example": "me", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all the substrates that a user made",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubstrateListResponseDto"
                }
              }
            }
          },
          "401": { "description": "Unauthorized - you need to be signed in" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Users"]
      }
    },
...
  },
  "info": {
    "title": "Cultivation App",
    "description": "Cultivation API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
      "SpeciesListResponseDto": { "type": "object", "properties": {} },
      "SpeciesResponseDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 255 },
          "id": {
            "type": "number",
            "example": 1,
            "description": "Unique identifier of the species"
          }
        },
        "required": ["name", "id"]
      },
      "CreateSpeciesRequestDto": {
        "type": "object",
        "properties": { "name": { "type": "string", "maxLength": 255 } },
        "required": ["name"]
      },
      "UpdateSpeciesRequestDto": {
        "type": "object",
        "properties": { "name": { "type": "string", "maxLength": 255 } },
        "required": ["name"]
      },
...
      "GrainSpawnResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Unique identifier of the grain spawn"
          },
          "inoculationDate": {
            "type": "string",
            "example": "2025-01-25T00:00:00.000Z",
            "description": "Date when grain spawn was inoculated",
            "format": "date-time"
          },
          "contaminationStatus": {
            "type": "boolean",
            "example": false,
            "description": "Contamination status of the grain spawn",
            "nullable": true
          },
          "shaken": {
            "type": "boolean",
            "example": true,
            "description": "Whether the grain spawn has been shaken",
            "nullable": true
          },
          "motherCultureId": {
            "type": "object",
            "example": 1,
            "description": "ID of the mother culture used",
            "nullable": true
          },
          "liquidCultureId": {
            "type": "object",
            "example": null,
            "description": "ID of the liquid culture used",
            "nullable": true
          },
          "species": { "$ref": "#/components/schemas/SpeciesResponseDto" }
        },
        "required": [
          "id",
          "inoculationDate",
          "contaminationStatus",
          "shaken",
          "motherCultureId",
          "liquidCultureId",
          "species"
        ]
      },
...
      "SubstrateResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Unique identifier of the substrate"
          },
          "inoculationDate": {
            "type": "string",
            "example": "2025-02-01T00:00:00.000Z",
            "description": "Date when substrate was inoculated",
            "format": "date-time"
          },
          "incubationDate": {
            "type": "string",
            "example": "2025-02-15T00:00:00.000Z",
            "description": "Date when substrate started incubation",
            "nullable": true,
            "format": "date-time"
          },
          "contaminationStatus": {
            "type": "boolean",
            "example": false,
            "description": "Contamination status of the substrate",
            "nullable": true
          },
          "grainSpawnId": {
            "type": "number",
            "example": 1,
            "description": "ID of the grain spawn used"
          }
        },
        "required": [
          "id",
          "inoculationDate",
          "incubationDate",
          "contaminationStatus",
          "grainSpawnId"
        ]
      },
      "SubstrateListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SubstrateResponseDto" }
          }
        },
        "required": ["items"]
      },
      "SubstrateDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Unique identifier of the substrate"
          },
          "inoculationDate": {
            "type": "string",
            "example": "2025-02-01T00:00:00.000Z",
            "description": "Date when substrate was inoculated",
            "format": "date-time"
          },
          "incubationDate": {
            "type": "string",
            "example": "2025-02-15T00:00:00.000Z",
            "description": "Date when substrate started incubation",
            "nullable": true,
            "format": "date-time"
          },
          "contaminationStatus": {
            "type": "boolean",
            "example": false,
            "description": "Contamination status of the substrate",
            "nullable": true
          },
          "grainSpawnId": {
            "type": "number",
            "example": 1,
            "description": "ID of the grain spawn used"
          },
          "grainSpawn": { "$ref": "#/components/schemas/GrainSpawnResponseDto" }
        },
        "required": [
          "id",
          "inoculationDate",
          "incubationDate",
          "contaminationStatus",
          "grainSpawnId",
          "grainSpawn"
        ]
      },
      "CreateSubstrateRequestDto": {
        "type": "object",
        "properties": {
          "grainSpawnId": { "type": "number", "minimum": 1 },
          "contaminationStatus": { "type": "boolean" }
        },
        "required": ["grainSpawnId", "contaminationStatus"]
      },
      "UpdateSubstrateRequestDto": { "type": "object", "properties": {} },
      "PublicUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "User ID",
            "minimum": 1,
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "User name",
            "minLength": 2,
            "maxLength": 255,
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@email.com",
            "format": "email"
          }
        },
        "required": ["id", "name", "email"]
      },
      "UserListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PublicUserResponseDto" }
          }
        },
        "required": ["items"]
      },
      "RegisterUserRequestDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 255 },
          "email": { "type": "string", "example": "user@email.com" },
          "password": { "type": "string", "minLength": 8, "maxLength": 128 }
        },
        "required": ["name", "email", "password"]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT token for authentication"
          }
        },
        "required": ["token"]
      },
      "UpdateUserRequestDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 255 },
          "email": { "type": "string", "example": "user@email.com" }
        },
        "required": ["name", "email"]
      },
...
    }
  }
}